Fixed SEGFAULT in grid_worker.cpp when -procinfo was not specified and the listening port was busy. JIRA: CXX-3813
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@57178 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/connect/services/grid_worker.cpp
@@ -1113,8 +1113,10 @@ int CGridWorkerNode::Run()
         control_thread->Prepare();
     }
     catch (CServer_Exception& e) {
-        fclose(procinfo_file);
-        unlink(procinfo_file_name.c_str());
+        if (procinfo_file != NULL) {
+            fclose(procinfo_file);
+            unlink(procinfo_file_name.c_str());
+        }
         if (e.GetErrCode() != CServer_Exception::eCouldntListen)
             throw;
         NCBI_THROW_FMT(CGridWorkerNodeException, ePortBusy,
